Refactor converter into modules and free functions
