+++ imessage-exporter/src/app/converters/convert.rs
@@ -1,0 +1,171 @@
+use std::{
+    fs::{copy, create_dir_all},
+    path::{Path, PathBuf},
+    process::{Command, Stdio},
+};
+
+use imessage_database::tables::attachment::MediaType;
+
+use crate::app::converters::models::{ImageConverter, ImageType};
+
+/// Copy a file without altering it
+pub(super) fn copy_raw(from: &Path, to: &Path) {
+    // Ensure the directory tree exists
+    if let Some(folder) = to.parent() {
+        if !folder.exists() {
+            if let Err(why) = create_dir_all(folder) {
+                eprintln!("Unable to create {folder:?}: {why}");
+            }
+        }
+    }
+    if let Err(why) = copy(from, to) {
+        eprintln!("Unable to copy {from:?} to {to:?}: {why}");
+    };
+}
+
+/// Copy amn image file, converting if possible
+///
+/// - Attachment `HEIC` files convert to `JPEG`
+/// - Fallback to the original format
+pub(super) fn image_copy_convert(
+    from: &Path,
+    to: &mut PathBuf,
+    converter: &ImageConverter,
+    mime_type: MediaType,
+) {
+    // Normal attachments always get converted to jpeg
+    if matches!(
+        mime_type,
+        MediaType::Image("heic") | MediaType::Image("HEIC")
+    ) {
+        let output_type = ImageType::Jpeg;
+        // Update extension for conversion
+        to.set_extension(output_type.to_str());
+        if convert_heic(from, to, converter, &output_type).is_none() {
+            eprintln!("Unable to convert {from:?}");
+        }
+    } else {
+        copy_raw(from, to);
+    }
+}
+
+/// Copy a sticker, converting if possible
+///
+/// - Sticker `HEIC` files convert to `PNG`
+/// - Sticker `HEICS` files convert to `GIF`
+/// - Fallback to the original format
+pub(super) fn sticker_copy_convert(
+    from: &Path,
+    to: &mut PathBuf,
+    converter: &ImageConverter,
+    mime_type: MediaType,
+) {
+    // Determine the output type of the sticker
+    let output_type: Option<ImageType> = match mime_type {
+        // Normal stickers get converted to png
+        MediaType::Image("heic") | MediaType::Image("HEIC") => Some(ImageType::Png),
+        MediaType::Image("heics")
+        | MediaType::Image("HEICS")
+        | MediaType::Image("heic-sequence") => Some(ImageType::Gif),
+        _ => None,
+    };
+
+    match output_type {
+        Some(output_type) => {
+            to.set_extension(output_type.to_str());
+            if convert_heic(from, to, converter, &output_type).is_none() {
+                eprintln!("Unable to convert {from:?}");
+            }
+        }
+        None => copy_raw(from, to),
+    }
+}
+
+/// Convert a HEIC image file to the provided format
+///
+/// This uses the macOS builtin `sips` program
+/// Docs: <https://www.unix.com/man-page/osx/1/sips/> (or `man sips`)
+///
+/// If `to` contains a directory that does not exist, i.e. `/fake/out.jpg`, instead
+/// of failing, `sips` will create a file called `fake` in `/`. Subsequent writes
+/// by `sips` to the same location will not fail, but since it is a file instead
+/// of a directory, this will fail for non-`sips` copies.
+pub(super) fn convert_heic(
+    from: &Path,
+    to: &Path,
+    converter: &ImageConverter,
+    output_image_type: &ImageType,
+) -> Option<()> {
+    // Get the path we want to copy from
+    let from_path = from.to_str()?;
+
+    // Get the path we want to write to
+    let to_path = to.to_str()?;
+
+    // Ensure the directory tree exists
+    if let Some(folder) = to.parent() {
+        if !folder.exists() {
+            if let Err(why) = create_dir_all(folder) {
+                eprintln!("Unable to create {folder:?}: {why}");
+                return None;
+            }
+        }
+    }
+
+    match converter {
+        ImageConverter::Sips => {
+            // Build the command
+            match Command::new("sips")
+                .args(vec![
+                    "-s",
+                    "format",
+                    output_image_type.to_str(),
+                    from_path,
+                    "-o",
+                    to_path,
+                ])
+                .stdout(Stdio::null())
+                .stderr(Stdio::null())
+                .stdin(Stdio::null())
+                .spawn()
+            {
+                Ok(mut sips) => match sips.wait() {
+                    Ok(_) => Some(()),
+                    Err(why) => {
+                        eprintln!("Conversion failed: {why}");
+                        None
+                    }
+                },
+                Err(why) => {
+                    eprintln!("Conversion failed: {why}");
+                    None
+                }
+            }
+        }
+        ImageConverter::Imagemagick =>
+        // Build the command
+        {
+            match Command::new("magick")
+                .args(vec![from_path, to_path])
+                .stdout(Stdio::null())
+                .stderr(Stdio::null())
+                .stdin(Stdio::null())
+                .spawn()
+            {
+                Ok(mut convert) => match convert.wait() {
+                    Ok(_) => Some(()),
+                    Err(why) => {
+                        eprintln!("Conversion failed: {why}");
+                        None
+                    }
+                },
+                Err(why) => {
+                    eprintln!("Conversion failed: {why}");
+                    None
+                }
+            }
+        }
+    };
+
+    Some(())
+}
