+++ imessage-exporter/src/app/mod.rs
@@ -1,7 +1,6 @@
-pub mod attachment_manager;
-pub mod converter;
 pub mod error;
 pub mod export_type;
+pub mod converters;
 pub mod options;
 pub mod progress;
 pub mod runtime;
